cardxml: Greatly simplify localized string loading
Use a common _locstring property for all LocStrings to make it
easier to change the logic.

Additionally, preparse all locstring locales into CardXML.tags.

This gets rid of the last remnants of lazy XML parsing, which was
mostly unnecessary as we were already doing it for enUS. It also
gets rid of CardXML._localized_tags.