Update player.starting_hero after Maestra of the Masquerade
The card "Maestra of the Masquerade" disguises the starting hero as
another hero. This leads to player.starting_hero pointing to a
misleading entity with the wrong player class. To fix this, we overwrite
the starting_hero property when the disguised hero is revealed to point
to the initially hidden Rogue hero.