Ensure Maestra of the Masquerade is in the known starting deck
This commit introduces the notion of well-known cards that are
guaranteed to be in the deck, even if the real entity has not been
revelead yet. This is initially used to ensure that Maestra of the
Masquerade is in the known starting deck after the hero reveal, but
could be expanded to support Start of Game effects in future.